+++ src/builtin/internal.rs
@@ -4,12 +4,11 @@ use crate::{
   Inst::*,
   Jsonpiler,
   LogicByteOpcode::*,
+  Memory::*,
   Operand::{Args, Ref},
   Register::*,
-  VarKind::*,
   utility::{mov_d, mov_q},
 };
-use std::collections::hash_map::Entry::{Occupied, Vacant};
 impl Jsonpiler {
   pub(crate) fn get_critical_section(&mut self) -> ErrOR<u32> {
     if let Some(id) = self.sym_table.get("CRITICAL_SECTION") {
@@ -23,82 +22,71 @@ impl Jsonpiler {
       LeaRM(Rcx, Global { id: critical_section, disp: 0i32 }),
       CallApi(initialize_critical_section),
     ]);
-    self.startup.extend_from_slice(&[
-      LeaRM(Rcx, Global { id: critical_section, disp: 0i32 }),
-      CallApi(initialize_critical_section),
-    ]);
     Ok(critical_section)
   }
   pub(crate) fn get_custom_error(&mut self, err_msg: &'static str) -> ErrOR<u32> {
+    if let Some(id) = self.sym_table.get(err_msg) {
+      return Ok(*id);
+    }
     let err_msg_id = self.global_str(err_msg.to_owned()).0;
     let message_box = self.import(Jsonpiler::USER32, "MessageBoxA")?;
-    let mb_a = self.call_api_check_null(message_box);
     let exit_process = self.import(Jsonpiler::KERNEL32, "ExitProcess")?;
-    match self.sym_table.entry(err_msg) {
-      Occupied(entry) => Ok(*entry.get()),
-      Vacant(entry) => {
-        let id = self.label_id;
-        self.label_id += 1;
-        self.insts.extend_from_slice(&[
-          Lbl(id),
-          Clear(Rcx),
-          LeaRM(Rdx, Global { id: err_msg_id, disp: 0i32 }),
-          Clear(R8),
-          mov_d(R9, 0x10),
-        ]);
-        self.insts.extend_from_slice(&mb_a);
-        self.insts.extend_from_slice(&[mov_d(Rcx, u32::MAX), CallApi(exit_process)]);
-        entry.insert(id);
-        Ok(id)
-      }
-    }
+    let id = self.gen_id();
+    self.sym_table.insert(err_msg, id);
+    self.insts.extend_from_slice(&[
+      Lbl(id),
+      Clear(Rcx),
+      LeaRM(Rdx, Global { id: err_msg_id, disp: 0i32 }),
+      Clear(R8),
+      mov_d(R9, 0x10),
+    ]);
+    self.insts.extend_from_slice(&self.call_api_check_null(message_box));
+    self.insts.extend_from_slice(&[mov_d(Rcx, u32::MAX), CallApi(exit_process)]);
+    Ok(id)
   }
   pub(crate) fn get_msg_box(&mut self) -> ErrOR<u32> {
+    if let Some(id) = self.sym_table.get("MSG_BOX") {
+      return Ok(*id);
+    }
     let heap = Global { id: self.sym_table["HEAP"], disp: 0i32 };
     let u8_to_16 = self.get_u8_to_16()?;
     let heap_free = self.import(Jsonpiler::KERNEL32, "HeapFree")?;
     let message_box_w = self.import(Jsonpiler::USER32, "MessageBoxW")?;
     let msg_box_insts = self.call_api_check_null(message_box_w);
     let heap_free_insts = self.call_api_check_null(heap_free);
-    match self.sym_table.entry("MSG_BOX") {
-      Occupied(entry) => Ok(*entry.get()),
-      Vacant(entry) => {
-        let id = self.label_id;
-        self.label_id += 1;
-        self.insts.extend_from_slice(&[
-          Lbl(id),
-          Push(Rdi),
-          Push(Rsi),
-          Push(Rbp),
-          mov_q(Rbp, Rsp),
-          SubRId(Rsp, 0x20),
-          mov_q(Rsi, Rdx),
-          Call(u8_to_16),
-          mov_q(Rdi, Rax),
-          mov_q(Rcx, Rsi),
-          Call(u8_to_16),
-          mov_q(Rsi, Rax),
-          Clear(Rcx),
-          mov_q(Rdx, Rsi),
-          mov_q(R8, Rdi),
-          Clear(R9),
-        ]);
-        self.insts.extend_from_slice(&msg_box_insts);
-        self.insts.extend_from_slice(&[mov_q(Rcx, heap), Clear(Rdx), mov_q(R8, Rdi)]);
-        self.insts.extend_from_slice(&heap_free_insts);
-        self.insts.extend_from_slice(&[mov_q(Rcx, heap), Clear(Rdx), mov_q(R8, Rsi)]);
-        self.insts.extend_from_slice(&heap_free_insts);
-        self.insts.extend_from_slice(&[
-          mov_q(Rsp, Rbp),
-          Pop(Rbp),
-          Pop(Rsi),
-          Pop(Rdi),
-          Custom(&Jsonpiler::RET),
-        ]);
-        entry.insert(id);
-        Ok(id)
-      }
-    }
+    let id = self.gen_id();
+    self.sym_table.insert("MSG_BOX", id);
+    self.insts.extend_from_slice(&[
+      Lbl(id),
+      Push(Rdi),
+      Push(Rsi),
+      Push(Rbp),
+      mov_q(Rbp, Rsp),
+      SubRId(Rsp, 0x20),
+      mov_q(Rsi, Rdx),
+      Call(u8_to_16),
+      mov_q(Rdi, Rax),
+      mov_q(Rcx, Rsi),
+      Call(u8_to_16),
+      mov_q(Rsi, Rax),
+      Clear(Rcx),
+      mov_q(Rdx, Rsi),
+      mov_q(R8, Rdi),
+      Clear(R9),
+    ]);
+    self.insts.extend_from_slice(&msg_box_insts);
+    self.insts.extend_from_slice(&[mov_q(Rcx, heap), Clear(Rdx), mov_q(R8, Rdi)]);
+    self.insts.extend_from_slice(&heap_free_insts);
+    self.insts.extend_from_slice(&[mov_q(Rcx, heap), Clear(Rdx), mov_q(R8, Rsi)]);
+    self.insts.extend_from_slice(&heap_free_insts);
+    self.insts.extend_from_slice(&[
+      mov_q(Rsp, Rbp),
+      Pop(Rbp),
+      Pop(Rsi),
+      Pop(Rdi),
+      Custom(&Jsonpiler::RET),
+    ]);
+    Ok(id)
   }
   pub(crate) fn get_random(&mut self) -> ErrOR<u32> {
     if let Some(id) = self.sym_table.get("RANDOM") {
@@ -134,65 +122,60 @@ impl Jsonpiler {
     Ok(id)
   }
   pub(crate) fn get_u8_to_16(&mut self) -> ErrOR<u32> {
+    if let Some(id) = self.sym_table.get("U8TO16") {
+      return Ok(*id);
+    }
     let heap = self.sym_table["HEAP"];
     let multi_byte_to_wide_char = self.import(Jsonpiler::KERNEL32, "MultiByteToWideChar")?;
-    let mb_t_wc = self.call_api_check_null(multi_byte_to_wide_char);
     let heap_alloc = self.import(Jsonpiler::KERNEL32, "HeapAlloc")?;
-    let h_a = self.call_api_check_null(heap_alloc);
-    match self.sym_table.entry("U8TO16") {
-      Occupied(entry) => Ok(*entry.get()),
-      Vacant(entry) => {
-        let id = self.label_id;
-        self.label_id += 1;
-        self.insts.extend_from_slice(&[
-          Lbl(id),
-          Push(Rdi),
-          Push(Rsi),
-          Push(Rbx),
-          Push(Rbp),
-          mov_q(Rbp, Rsp),
-          SubRId(Rsp, 0x48),
-          mov_q(Rdi, Rcx),
-          mov_d(Rcx, 65001),
-          Clear(Rdx),
-          mov_q(R8, Rdi),
-          mov_d(R9, u32::MAX),
-          Clear(Rax),
-          mov_q(Args(0x20), Rax),
-          mov_q(Args(0x28), Rax),
-        ]);
-        self.insts.extend_from_slice(&mb_t_wc);
-        self.insts.extend_from_slice(&[
-          Shl1R(Rax),
-          mov_q(Rsi, Rax),
-          mov_q(Rcx, Global { id: heap, disp: 0i32 }),
-          Clear(Rdx),
-          mov_q(R8, Rsi),
-        ]);
-        self.insts.extend_from_slice(&h_a);
-        self.insts.extend_from_slice(&[
-          mov_q(Rbx, Rax),
-          mov_d(Rcx, 65001),
-          Clear(Rdx),
-          mov_q(R8, Rdi),
-          mov_d(R9, u32::MAX),
-          mov_q(Args(0x20), Rbx),
-          mov_q(Args(0x28), Rsi),
-        ]);
-        self.insts.extend_from_slice(&mb_t_wc);
-        self.insts.extend_from_slice(&[
-          mov_q(Rax, Rbx),
-          mov_q(Rsp, Rbp),
-          Pop(Rbp),
-          Pop(Rbx),
-          Pop(Rsi),
-          Pop(Rdi),
-          Custom(&Jsonpiler::RET),
-        ]);
-        entry.insert(id);
-        Ok(id)
-      }
-    }
+    let id = self.gen_id();
+    self.sym_table.insert("U8TO16", id);
+    self.insts.extend_from_slice(&[
+      Lbl(id),
+      Push(Rdi),
+      Push(Rsi),
+      Push(Rbx),
+      Push(Rbp),
+      mov_q(Rbp, Rsp),
+      SubRId(Rsp, 0x48),
+      mov_q(Rdi, Rcx),
+      mov_d(Rcx, 65001),
+      Clear(Rdx),
+      mov_q(R8, Rdi),
+      mov_d(R9, u32::MAX),
+      Clear(Rax),
+      mov_q(Args(0x20), Rax),
+      mov_q(Args(0x28), Rax),
+    ]);
+    self.insts.extend_from_slice(&self.call_api_check_null(multi_byte_to_wide_char));
+    self.insts.extend_from_slice(&[
+      Shl1R(Rax),
+      mov_q(Rsi, Rax),
+      mov_q(Rcx, Global { id: heap, disp: 0i32 }),
+      Clear(Rdx),
+      mov_q(R8, Rsi),
+    ]);
+    self.insts.extend_from_slice(&self.call_api_check_null(heap_alloc));
+    self.insts.extend_from_slice(&[
+      mov_q(Rbx, Rax),
+      mov_d(Rcx, 65001),
+      Clear(Rdx),
+      mov_q(R8, Rdi),
+      mov_d(R9, u32::MAX),
+      mov_q(Args(0x20), Rbx),
+      mov_q(Args(0x28), Rsi),
+    ]);
+    self.insts.extend_from_slice(&self.call_api_check_null(multi_byte_to_wide_char));
+    self.insts.extend_from_slice(&[
+      mov_q(Rax, Rbx),
+      mov_q(Rsp, Rbp),
+      Pop(Rbp),
+      Pop(Rbx),
+      Pop(Rsi),
+      Pop(Rdi),
+      Custom(&Jsonpiler::RET),
+    ]);
+    Ok(id)
   }
   #[expect(clippy::too_many_lines)]
   pub(crate) fn get_wnd_proc(&mut self, pixel_func: u32) -> ErrOR<u32> {
@@ -210,9 +193,7 @@ impl Jsonpiler {
     let get_client_rect = self.import(Jsonpiler::USER32, "GetClientRect")?;
     let stretch_di_bits = self.import(Jsonpiler::GDI32, "StretchDIBits")?;
     let gui_frame = self.get_bss_id(8, 8);
-    self.sym_table.insert("GUI_TICK", gui_frame);
     let hwnd = self.get_bss_id(8, 8);
-    self.sym_table.insert("GUI_HWND", hwnd);
     let heap = self.sym_table["HEAP"];
     let gui_pixels = self.get_bss_id(8, 8);
     let bm_info = self.get_bss_id(44, 8);
@@ -239,14 +220,14 @@ impl Jsonpiler {
       SubRId(Rsp, 0x90),
       mov_q(Global { id: hwnd, disp: 0i32 }, Rcx),
       CmpRIb(Rdx, 0x2),
-      Jcc(E, handle_wm_destroy),
+      JCc(E, handle_wm_destroy),
       CmpRIb(Rdx, 0xf),
-      Jcc(E, handle_wm_paint),
+      JCc(E, handle_wm_paint),
       CmpRIb(Rdx, 0x1),
-      Jcc(E, handle_wm_create),
+      JCc(E, handle_wm_create),
       mov_d(Rax, 0x113),
       LogicRR(Cmp, Rdx, Rax),
-      Jcc(E, handle_wm_timer),
+      JCc(E, handle_wm_timer),
       CallApi(def_window_proc),
       Jmp(end_wnd_proc),
       Lbl(handle_wm_destroy),
@@ -287,7 +268,7 @@ impl Jsonpiler {
       Clear(Rax),
       Jmp(end_wnd_proc),
       Lbl(handle_wm_paint),
-      LeaRM(Rcx, Local { offset: 16i32 }),
+      LeaRM(Rcx, Local { offset: 0x10i32 }),
     ]);
     self.insts.extend_from_slice(&self.call_api_check_null(get_cursor_pos));
     self.insts.extend_from_slice(&[
@@ -307,8 +288,9 @@ impl Jsonpiler {
       mov_d(Rcx, Local { offset: 0x18i32 }),
       mov_d(Rdx, Local { offset: 0x20i32 }),
       SubRR(Rcx, Rdx),
-      LogicRR(Test, Rcx, Rcx),
-      Jcc(E, idiv_zero_w),
+      Clear(Rdx),
+      LogicRR(Cmp, Rcx, Rdx),
+      JCc(Le, idiv_zero_h),
       Custom(&Jsonpiler::CQO),
       IDivR(Rcx),
       Jmp(idiv_end_w),
@@ -322,8 +304,9 @@ impl Jsonpiler {
       mov_d(Rcx, Local { offset: 0x14i32 }),
       mov_d(Rdx, Local { offset: 0x1Ci32 }),
       SubRR(Rcx, Rdx),
-      LogicRR(Test, Rcx, Rcx),
-      Jcc(E, idiv_zero_h),
+      Clear(Rdx),
+      LogicRR(Cmp, Rcx, Rdx),
+      JCc(Le, idiv_zero_h),
       Custom(&Jsonpiler::CQO),
       IDivR(Rcx),
       Jmp(idiv_end_h),
@@ -336,24 +319,30 @@ impl Jsonpiler {
       mov_d(Rcx, Local { offset: 4i32 }),
       mov_d(Rdx, Jsonpiler::GUI_H),
       LogicRR(Cmp, Rcx, Rdx),
-      Jcc(E, while_end_y),
+      JCc(E, while_end_y),
       mov_d(Local { offset: 8i32 }, 0),
       Lbl(while_x),
       mov_d(Rcx, Local { offset: 8i32 }),
       mov_d(Rdx, Jsonpiler::GUI_W),
       LogicRR(Cmp, Rcx, Rdx),
-      Jcc(E, while_end_x),
+      JCc(E, while_end_x),
       mov_d(Rcx, Local { offset: 8i32 }),
       SubRId(Rcx, Jsonpiler::GUI_W >> 1),
       mov_d(Rdx, Local { offset: 4i32 }),
       SubRId(Rdx, Jsonpiler::GUI_H >> 1),
       mov_q(R8, Global { id: gui_frame, disp: 0i32 }),
-      mov_d(Rax, Jsonpiler::GUI_H >> 1u8),
       mov_d(R9, Local { offset: 0xCi32 }),
-      SubRR(Rax, R9),
-      mov_q(Args(0x20), Rax),
-      mov_d(R9, Local { offset: 0x10i32 }),
+      mov_d(Rax, Jsonpiler::GUI_W),
+      LogicRR(Cmp, R9, Rax),
+      CMovCc(G, R9, Rax),
       SubRId(R9, Jsonpiler::GUI_W >> 1),
+      NegR(R9),
+      mov_q(Args(0x20), R9),
+      mov_d(R9, Local { offset: 0x10i32 }),
+      mov_d(Rax, Jsonpiler::GUI_H),
+      LogicRR(Cmp, R9, Rax),
+      CMovCc(G, R9, Rax),
+      SubRId(R9, Jsonpiler::GUI_H >> 1),
       Call(pixel_func),
       mov_d(Rcx, Local { offset: 4i32 }),
       mov_d(Rdx, Jsonpiler::GUI_W),